refactor: Creating DTO User.
